src/auth/auth.controller.ts
import { Body, Controller, Headers, Post } from '@nestjs/common';
import { AuthService } from './auth.service';

@Controller('auth')
export class AuthController {
  constructor(private authService: AuthService) {}

  @Post('signup')
  signUp(@Body() body: { email: string; password: string }) {
    return this.authService.signUp(body.email, body.password);
  }

  @Post('signin')
  signIn(@Body() body: { email: string; password: string }) {
    return this.authService.signIn(body.email, body.password);
  }

  @Post('challange')
  recycleAccessToken(@Headers() headers: { authorization: string }) {
    const accessToken = headers['authorization'];

    return this.authService.verifyAccessToken(accessToken);
  }
}
